refactor: Move generation utility functions to separate module
